ElastiCache-Redis/cdk/cache/requirements.txt
aws-cdk-lib==2.158.0
constructs==10.3.0
